+++ screens/auth/stack.js
@@ -2,7 +2,7 @@ import * as React from "react";
 import { createStackNavigator } from "@react-navigation/stack";
 import SignInScreen from "./SignInScreen";
 import SignUpScreen from "./SignUpScreen";
-import RecoverPassword from "./RecoverPassword";
+
 const AuthStack = createStackNavigator();
 
 export default function AuthStackScreen() {
@@ -17,11 +17,6 @@ export default function AuthStackScreen() {
         name="SignUp"
         component={SignUpScreen}
         options={{ title: "Sign up" }}
-      />
-      <AuthStack.Screen
-        name="RecoverPassword"
-        component={RecoverPassword}
-        options={{ title: "Recover Password" }}
       />
     </AuthStack.Navigator>
   );
